requirements.txt
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.2.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-filter==23.1
django-login-required-middleware==0.9.0
django-querysetsequence==0.16
executing==1.2.0
fastjsonschema==2.20.0
feedparser==6.0.10
fonttools==4.53.1
fqdn==1.5.1
gevent==24.10.3
greenlet==3.1.1
h11==0.14.0
HackerNews==2.0.0
hackernews-python==0.3.2
httpcore==1.0.5
httpx==0.27.2
idna==3.4
imageio==2.35.1
imbalanced-learn==0.12.3
ipykernel==6.29.5
ipython==8.13.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
kiwisolver==1.4.7
lazy_loader==0.4
lime==0.2.0.1
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.0.0
oauthlib==3.2.2
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.2
praw==7.7.0
prawcore==2.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.38
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.15.1
pyparsing==3.1.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.17.0
setuptools==72.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==2.0.34
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.9.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.0.2
wcwidth==0.2.6
webcolors==24.8.0
webencodings==0.5.1
wheel==0.43.0
widgetsnbextension==4.0.13
zope.event==5.0
zope.interface==7.1.1
